renamed Core -> Tools
