+++ bin/BuildFile.xml
@@ -6,7 +6,7 @@
 <use name="roofit" />
 <use name="boost" />
 <use name="boost_program_options" />
-<use name="Analysis/Core" />
+<use name="Analysis/Tools" />
 
 <bin   name="AnalysisEventInfo" file="AnalysisEventInfo.cc">
 <flags   LDFLAGS="-lCore -lRIO -lNet -lHist -lGraf -lGraf3d -lGpad -lTree -lRint -lPostscript -lMatrix -lPhysics -lMathCore -lThread -lz -pthread -lm -ldl -rdynamic"/>
