+++ bin/ExampleToolsForNtuple.cc
@@ -7,12 +7,12 @@
 #include "TChain.h"
 #include "TH1.h" 
 
-#include "Analysis/Core/interface/JetCollection.h"
-#include "Analysis/Core/interface/Muons.h"
+#include "Analysis/Tools/interface/JetCollection.h"
+#include "Analysis/Tools/interface/Muons.h"
 
 using namespace std;
 using namespace analysis;
-using namespace analysis::core;
+using namespace analysis::tools;
 
 
 void PrintCrossSections(TChain *);
