+++ bin/SignalEffTrigger.cc
@@ -8,11 +8,11 @@
 #include "TH1.h" 
 #include "TGraphAsymmErrors.h" 
 
-#include "Analysis/Core/interface/Analysis.h"
+#include "Analysis/Tools/interface/Analysis.h"
 
 using namespace std;
 using namespace analysis;
-using namespace analysis::core;
+using namespace analysis::tools;
 
 
 // =============================================================================================
