+++ interface/Collection.h
@@ -1,12 +1,12 @@
-#ifndef Analysis_Core_Collection_h
-#define Analysis_Core_Collection_h 1
+#ifndef Analysis_Tools_Collection_h
+#define Analysis_Tools_Collection_h 1
 
 // -*- C++ -*-
 //
-// Package:    Analysis/Core
+// Package:    Analysis/Tools
 // Class:      Collection
 // 
-/**\class Collection Collection.cc Analysis/Core/src/Collection.cc
+/**\class Collection Collection.cc Analysis/Tools/src/Collection.cc
 
  Description: [one line class summary]
 
@@ -24,17 +24,17 @@
 #include <vector>
 // 
 // user include files
-#include "Analysis/Core/interface/Candidate.h"
-#include "Analysis/Core/interface/Jet.h"
-#include "Analysis/Core/interface/TriggerObject.h"
-#include "Analysis/Core/interface/GenParticle.h"
+#include "Analysis/Tools/interface/Candidate.h"
+#include "Analysis/Tools/interface/Jet.h"
+#include "Analysis/Tools/interface/TriggerObject.h"
+#include "Analysis/Tools/interface/GenParticle.h"
 
 //
 // class declaration
 //
 
 namespace analysis {
-   namespace core {
+   namespace tools {
       
       template <class Object>
       class Collection {
@@ -93,4 +93,4 @@ namespace analysis {
    }
 }
 
-#endif  // Analysis_Core_Tree_h
+#endif  // Analysis_Tools_Tree_h
