+++ src/utilLib.cpp
@@ -1,4 +1,4 @@
-#include "Analysis/Core/interface/utilLib.h"
+#include "Analysis/Tools/interface/utilLib.h"
 
 bool findStrings(const std::string & input, const std::string & needful)
 {
