+++ test/submitCrab3.py
@@ -27,7 +27,7 @@ if __name__ == '__main__':
    from CRABClient.ClientExceptions import ClientException
    from httplib import HTTPException
     
-   from Analysis.Core.crabConfig import crabConfig
+   from Analysis.Tools.crabConfig import crabConfig
    config = crabConfig()
    
    if TYPE == 'MC':
